requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.15.0
folium>=0.14.0
streamlit-folium>=0.15.0
altair>=5.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Natural Language Processing & Search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
transformers>=4.30.0

# Data Processing
openpyxl>=3.1.0
xlrd>=2.0.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pathlib>=1.0.1

# Development Dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional Performance Enhancements
watchdog>=3.0.0
